+++ src/main/java/de/tub/ise/anwsys/controllers/PizzaController.java
@@ -2,11 +2,14 @@ package de.tub.ise.anwsys.controllers;
 
 import de.tub.ise.anwsys.CustomExceptions.InvalidInputException;
 import de.tub.ise.anwsys.CustomExceptions.ItemNotFoundException;
+import de.tub.ise.anwsys.CustomStatus.CustomHttpResponse;
 import de.tub.ise.anwsys.models.Pizza;
 import de.tub.ise.anwsys.repos.PizzaRepository;
 import org.h2.util.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.orm.jpa.JpaObjectRetrievalFailureException;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
@@ -23,13 +26,13 @@ public class PizzaController {
 	private PizzaRepository pizzaRepository;
 
 	@RequestMapping("/pizza")
-	public List<Long> getAllPizzas() throws ItemNotFoundException {
+	public ResponseEntity<List<Long>> getAllPizzas() throws ItemNotFoundException {
 		ArrayList<Long> pizzas = new ArrayList<>();
 		pizzaRepository.findAll().forEach(x -> pizzas.add(x.getId()));
 		if (pizzas.size() == 0) {
 			throw new ItemNotFoundException("No pizzas exist");
 		}
-		return pizzas;
+		return CustomHttpResponse.createResponse(HttpStatus.OK, pizzas);
 	}
 
 	@RequestMapping(value = "/pizza/{pizzaId}", method = RequestMethod.GET)
@@ -43,9 +46,14 @@ public class PizzaController {
 
 
 	@RequestMapping(value = "/pizza", method = RequestMethod.POST)
-	public String addPizza(@RequestBody Pizza pizza) throws InvalidInputException {
-		if (pizza == null || StringUtils.isNullOrEmpty(pizza.getName())) {
+	public ResponseEntity<String> addPizza(@RequestBody Pizza pizza) throws InvalidInputException {
+
+		if (pizza == null) {
 			throw new InvalidInputException("Invalid input");
+		}
+		String name = pizza.getName() != null ? pizza.getName().trim() : "";
+		if (StringUtils.isNullOrEmpty(name)) {
+			throw new InvalidInputException("Name is required");
 		}
 		if (StringUtils.isNullOrEmpty(pizza.getSize())) {
 			pizza.setSize(Pizza.Size.Standard.name());
@@ -56,43 +64,52 @@ public class PizzaController {
 			}
 		}
 		try {
-			pizzaRepository.save(pizza);
+			pizza.setId(0);
+			pizza = pizzaRepository.save(pizza);
 		} catch (DataIntegrityViolationException ex) {
 			throw new InvalidInputException("Another Pizza with the same name exists");
 		}
 		URI location = ServletUriComponentsBuilder
 				.fromCurrentRequest().path("/{pizzaId}")
 				.buildAndExpand(pizza.getId()).toUri();
-		return location.getRawPath();
+		return CustomHttpResponse.createResponse(HttpStatus.CREATED, "location", location.toString(), "Created new pizza.");
 	}
 
 	@RequestMapping(value = "/pizza/{pizzaId}", method = RequestMethod.PUT)
-	public String updatePizza(@RequestBody Pizza pizza, @PathVariable Long pizzaId) throws ItemNotFoundException {
+	public ResponseEntity<String> updatePizza(@RequestBody Pizza pizza, @PathVariable Long pizzaId) throws ItemNotFoundException {
+
+		pizza.setId(pizzaId);
 		if (pizza == null) {
 			throw new InvalidInputException("Invalid pizza supplied");
 		}
-		pizza.setId(pizzaId);
-		Pizza.Size size = Pizza.Size.parse(pizza.getSize());
-		if (size == null) {
-			throw new InvalidInputException("The size is invalid. either enter Large or Standard");
+		if (StringUtils.isNullOrEmpty(pizza.getSize())) {
+			pizza.setSize(Pizza.Size.Standard.name());
+		} else {
+			Pizza.Size size = Pizza.Size.parse(pizza.getSize());
+			if (size == null) {
+				throw new InvalidInputException("The size is invalid. either enter Large or Standard");
+			}
+		}
+		if (pizzaRepository.findOne(pizzaId) == null) {
+			throw new ItemNotFoundException("Pizza not found");
 		}
 		try {
-			pizzaRepository.save(pizza);
+			pizza = pizzaRepository.save(pizza);
 		} catch (DataIntegrityViolationException ex) {
 			throw new InvalidInputException("Another Pizza with the same name exists in the szstem");
 		}
-		return "Update okay";
+		return CustomHttpResponse.createResponse(HttpStatus.NO_CONTENT, "Update okay");
 	}
 
 
 	@RequestMapping(value = "/pizza/{pizzaId}", method = RequestMethod.DELETE)
-	public String deletePizza(@PathVariable Long pizzaId) throws ItemNotFoundException, InvalidInputException {
-//		if (pizzaRepository.findOne(pizzaId) == null) {
-//			throw new ItemNotFoundException("Pizza not found");
-//		}
+	public ResponseEntity<String> deletePizza(@PathVariable Long pizzaId) throws ItemNotFoundException, InvalidInputException {
+		if (pizzaRepository.findOne(pizzaId) == null) {
+			throw new ItemNotFoundException("Pizza not found exception");
+		}
 		try {
 			pizzaRepository.delete(pizzaId);
-			return "Deleted";
+			return CustomHttpResponse.createResponse(HttpStatus.NO_CONTENT, "Deleted");
 		} catch (JpaObjectRetrievalFailureException ex) {
 			throw new ItemNotFoundException("The order with the specified id might not be available");
 		} catch (DataIntegrityViolationException ex) {
